+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.16] (2022-04-25)
+
 ### Added
 
 - **publish:** Check if a publication can be configured automatically for the project
@@ -496,6 +498,7 @@ Pull request: #35
 - Added CHANGELOG.md :)
 
 [unreleased]: https://github.com/RedMadRobot/gradle-infrastructure/compare/main..develop
+[0.16]: https://github.com/RedMadRobot/gradle-infrastructure/compare/v0.15..v0.16
 [0.15]: https://github.com/RedMadRobot/gradle-infrastructure/compare/v0.14..v0.15
 [0.14]: https://github.com/RedMadRobot/gradle-infrastructure/compare/v0.13..v0.14
 [0.13]: https://github.com/RedMadRobot/gradle-infrastructure/compare/v0.12.2...v0.13